Thoughts and extra points on RFM list
I have gone through the entire list of RFMs and have added thoughts/questions, or extended the methodology in some of them. I would like to do another pass to get into more detail (at least for some of them), but would like to open the PR for discussion before I do that.